+++ CHANGELOG.md
@@ -11,5 +11,6 @@
 ### Fixed
 
 - Fixed bug where an opened collapsed tab would close when clicking the container without clicking a list item.
+- Fixed issue with disabled loading button not displaying properly (#61).
 
 ### Removed